--- Assets/Prefabs/Client.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 2f89945cea7b9ec44a601866f29a32c9
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
