--- Assets/Prefabs/Default/PlayerStats.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 55d63129ef1fd7d4ca9e0acb7141d022
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
