--- Assets/Scripts/Editor.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5a2e144eb6c477743af7dec12d434c7a
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
