+++ Assets/Scripts/HealthSystem/HealthBar.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 33a6a679984a965479e1d38aaef7ae67
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
